Create README.md
Tic-Tac-Toe

Tic Tac Toe. This is a game for two people. Each person will be represented with X’s and O’s. You have a 3 by 3 grid in which two players place x or o. Three characters in a row wins! Rules of the game. Each person will take a turn to mark any of a spot that is open on the bord. To win the game, a player needs to occupy 3 consecutive positions in any direction on the board.

Plans to represent in a console program. 1.Find a way to represent the Tic Tac Toe table with data structure. 2. Create player one attributes 3.Create player two attributes 4.Find a way to interact with players. 5.Be able to read input from player 6. Validate players input 7. Palace players input at an appropriate location on the table 8. Read the table and assess if there is a winner. If not, continue the game. 9.Anounce the name of winner when winning rules satisfied